+++ pictureTape/Auth/OAuth2Service.swift
@@ -8,38 +8,38 @@
 import Foundation
 
 
-    final class OAuth2Service {
-        static let shared = OAuth2Service()
-        private var splashViewController = SplashViewController()
-        private let urlSession = URLSession.shared
-        private (set) var authToken: String {
-            get {
-                return OAuth2TokenStorage().token!
-            }
-            set {
-                OAuth2TokenStorage().token = newValue
-                print("TOKEN", newValue)
-    } }
-        func fetchOAuthToken(
-          _ code: String,
-          completion: @escaping (Result<String, Error>) -> Void
-    ){
-    let request = authTokenRequest(code: code)
-    let task = object(for: request) { [weak self] result in
-                guard let self = self else { return }
-                switch result {
-                case .success(let body):
-                    let authToken = body.accessToken
-                    self.authToken = authToken
-                    completion(.success(authToken))
-                    splashViewController.switchToTabBarController()
-                case .failure(let error):
-                    completion(.failure(error))
-    } }
-            
-            task.resume()
+final class OAuth2Service {
+    static let shared = OAuth2Service()
+    private var splashViewController = SplashViewController()
+    private let urlSession = URLSession.shared
+    private (set) var authToken: String {
+        get {
+            return OAuth2TokenStorage().token!
         }
+        set {
+            OAuth2TokenStorage().token = newValue
+            print("TOKEN", newValue)
+        } }
+    func fetchOAuthToken(
+        _ code: String,
+        completion: @escaping (Result<String, Error>) -> Void
+    ){
+        let request = authTokenRequest(code: code)
+        let task = object(for: request) { [weak self] result in
+            guard let self = self else { return }
+            switch result {
+            case .success(let body):
+                let authToken = body.accessToken
+                self.authToken = authToken
+                completion(.success(authToken))
+                splashViewController.switchToTabBarController()
+            case .failure(let error):
+                completion(.failure(error))
+            } }
+        
+        task.resume()
     }
+}
 extension OAuth2Service {
     private func object(
         for request: URLRequest,
@@ -49,66 +49,21 @@ extension OAuth2Service {
         return urlSession.data(for: request) { (result: Result<Data, Error>) in
             let response = result.flatMap { data -> Result<OAuthTokenResponseBody, Error> in
                 Result {
-                  try decoder.decode(OAuthTokenResponseBody.self, from: data)
+                    try decoder.decode(OAuthTokenResponseBody.self, from: data)
                 }
-}
+            }
             completion(response)
         }
-} }
+    } }
 private func authTokenRequest(code: String) -> URLRequest {
-      URLRequest.makeHTTPRequest(
-          path: "/oauth/token"
-          + "?client_id=\(accessKey)"
-          + "&&client_secret=\(secretKey)"
-          + "&&redirect_uri=\(redirectURI)"
-          + "&&code=\(code)"
-          + "&&grant_type=authorization_code",
-          httpMethod: "POST",
-          baseURL: URL(string: "https://unsplash.com")!
-) }
+    URLRequest.makeHTTPRequest(
+        path: "/oauth/token"
+        + "?client_id=\(accessKey)"
+        + "&&client_secret=\(secretKey)"
+        + "&&redirect_uri=\(redirectURI)"
+        + "&&code=\(code)"
+        + "&&grant_type=authorization_code",
+        httpMethod: "POST",
+        baseURL: URL(string: "https://unsplash.com")!
+    ) }
 
-enum NetworkError: Error {
-    case httpStatusCode(Int)
-    case urlRequestError(Error)
-    case urlSessionError
-}
-extension URLSession {
-    func data(
-        for request: URLRequest,
-        completion: @escaping (Result<Data, Error>) -> Void
-    ) -> URLSessionTask {
-        let fulfillCompletion: (Result<Data, Error>) -> Void = { result in
-            DispatchQueue.main.async {
-                completion(result)
-            }
-}
-        let task = dataTask(with: request) { data, response, error in
-            if let data = data,
-                let response = response,
-                let statusCode = (response as? HTTPURLResponse)?.statusCode
-            {
-                if 200 ..< 300 ~= statusCode {
-                    fulfillCompletion(.success(data))
-                } else {
-                    fulfillCompletion(.failure(NetworkError.httpStatusCode(statusCode)))
-                }
-            } else if let error = error {
-                fulfillCompletion(.failure(NetworkError.urlRequestError(error)))
-            } else {
-                fulfillCompletion(.failure(NetworkError.urlSessionError))
-            }
-        }
-        task.resume()
-        return task
-} }
-
-extension URLRequest {
-    static func makeHTTPRequest(
-        path: String,
-        httpMethod: String,
-        baseURL: URL = defaultBaseURL!
-    ) -> URLRequest {
-        var request = URLRequest(url: URL(string: path, relativeTo: baseURL)!)
-        request.httpMethod = httpMethod
-        return request
-} }
