+++ pictureTape/Auth/OAuth2TokenStorage.swift
@@ -9,14 +9,14 @@ import Foundation
 
 class OAuth2TokenStorage {
     private let tokenKey = "OAuth2AccessToken"
-       
+    
     var token: String? {
-           get {
-               return UserDefaults.standard.string(forKey: tokenKey)
-           }
-           set {
-               UserDefaults.standard.set(newValue, forKey: tokenKey)
-           }
-       }
+        get {
+            return UserDefaults.standard.string(forKey: tokenKey)
+        }
+        set {
+            UserDefaults.standard.set(newValue, forKey: tokenKey)
+        }
+    }
 }
- 
+
